[Feature] Override resource parse to cache meta
Our JsonApiResource class just returns meta as an array.
Therefore we need to determine if the meta is empty, while
caching the value so that the JsonApiResource meta methods
are not called twice. (That would not be ideal because the
developer could be doing expensive calculations to generate
meta.)

This is implemented by extending the Neomerx `IdentiferAndResource`
parser class. The extended class just gets meta from the schema
and caches the value. It uses the cached value to determine if
the meta is empty or not.