[Bugfix] Safely handle missing relationship links
The relationship document now does not assume the relationship
exists on the resource, as some developers are hiding the relation
meaning the relationship document would cause an error.

Also, it was previously assuming that both the self and the related
links existed on the relationship. However, one or both might not
exist. It now only merges relationship links that exist.

This change also means that all relationship links are now merged,
whereas before only the self and related links were merged.