[Refactor] Remove merging of relationship links
The core package now takes care of merging relationship links
in the relationship response classes, while providing the
ability to turn off the merging - which is a better implementation.

The removal of this merging is not considered breaking because
the core package dependency has been upgraded plus removing
the feature fixes bugs in the implementation within this package.
The bugs were the merging would fail if the relationship was
marked as hidden, or if either the self or related link did not
exist on the relationship.